[AO] new feature: add custom headers field to the rest query interface

+++ app/uk/gov/hmrc/agentsexternalstubsfrontend/controllers/RestQueryController.scala
@@ -29,10 +29,8 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.libs.json._
 import play.api.libs.ws.{WSClient, WSResponse}
 import play.api.mvc._
-import play.mvc.Http.HeaderNames
 import uk.gov.hmrc.agentsexternalstubsfrontend.connectors.AgentsExternalStubsConnector
 import uk.gov.hmrc.agentsexternalstubsfrontend.services.Features
-import uk.gov.hmrc.agentsexternalstubsfrontend.views.html
 import uk.gov.hmrc.agentsexternalstubsfrontend.views.html.rest_query
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.http.SessionKeys
@@ -40,6 +38,8 @@ import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
+import scala.util.Try
+import java.net.URL
 
 @Singleton
 class RestQueryController @Inject()(
@@ -65,7 +65,7 @@ class RestQueryController @Inject()(
             case None =>
               Future.successful(
                 Ok(restQueryView(
-                  RestQueryForm.fill(RestQuery("GET", "https://", None)),
+                  RestQueryForm.fill(RestQuery("GET", "https://", None, None)),
                   routes.RestQueryController.runQuery(),
                   routes.UserController.showUserPage(),
                   routes.RestQueryController.showRestQueryPage(None),
@@ -134,17 +134,24 @@ class RestQueryController @Inject()(
 
   private def runQuery(
     query: RestQuery)(implicit ec: ExecutionContext, request: Request[AnyContent]): Future[WSResponse] = {
-    val wsRequest = wsClient
-      .url(query.url)
-      .withHeaders(
-        HeaderNames.AUTHORIZATION -> request.session.get(SessionKeys.authToken).get,
-        "X-Session-ID"            -> request.session.get(SessionKeys.sessionId).get)
-    query.method match {
-      case "GET"    => wsRequest.get()
-      case "POST"   => wsRequest.post(query.payload.getOrElse(JsNull))
-      case "PUT"    => wsRequest.put(query.payload.getOrElse(JsNull))
-      case "DELETE" => wsRequest.delete()
-      case _        => Future.failed(new Exception(s"Method ${query.method} is not supported, try GET, POST, PUT or DELETE"))
+    if (Try(new URL(query.url)).isFailure) 
+      Future.failed(new Exception(s"Invalid URL ${query.url}"))
+    else {
+      Try{
+          val wsRequest = wsClient
+          .url(query.url)
+          .withHttpHeaders(query.headersWithDefault.toSeq:_*)
+        query.method match {
+          case "GET"    => wsRequest.get()
+          case "POST"   => wsRequest.post(query.payload.getOrElse(JsNull))
+          case "PUT"    => wsRequest.put(query.payload.getOrElse(JsNull))
+          case "DELETE" => wsRequest.delete()
+          case _        => Future.failed(new Exception(s"Method ${query.method} is not supported, try GET, POST, PUT or DELETE"))
+        }
+      }.fold(
+        e => Future.failed(new Exception(s"Error executing the request: ${e.getMessage}")),
+        identity
+      )
     }
   }
 
@@ -152,20 +159,25 @@ class RestQueryController @Inject()(
 
 object RestQueryController {
 
-  case class RestQuery(method: String, url: String, payload: Option[JsValue]) {
+  case class RestQuery(method: String, url: String, payload: Option[JsValue], headers: Option[String] = None) {
     def encode: String =
       new String(
         Base64.getUrlEncoder.encode(Json.toJson(this).toString().getBytes(StandardCharsets.UTF_8)),
         StandardCharsets.UTF_8)
 
-    def toCurlCommand(implicit request: Request[AnyContent]): String =
-      s"""curl -v -X $method -H "Authorization: ${request.session
-        .get(SessionKeys.authToken)
-        .get}" -H "X-Session-ID: ${request.session
-        .get(SessionKeys.sessionId)
-        .get}" ${if (method == "POST" || method == "PUT") """-H "Content-Type: application/json"""" else ""} ${payload
+    def headersWithDefault(implicit request: Request[AnyContent]): Map[String,String] = {
+      val h = headers.map(RestQueryController.parseHeaders).getOrElse(Map.empty)
+      h ++ 
+      (if(h.exists(_._1.toLowerCase() == "authorization")) Map.empty else Map("Authorization" -> request.session.get(SessionKeys.authToken).get)) ++ 
+      (if(h.exists(_._1.toLowerCase() == "x-session-id")) Map.empty else Map("X-Session-ID" -> request.session.get(SessionKeys.sessionId).get)) ++
+      (if (!h.exists(_._1.toLowerCase() == "content-type") && (method == "POST" || method == "PUT")) Map("Content-Type" -> "application/json") else Map.empty)
+    }    
+
+    def toCurlCommand(implicit request: Request[AnyContent]): String = {
+      s"""curl -v -X $method ${headersWithDefault.map{case (key,value) => s"""-H "$key: $value""""}.mkString(" ")} ${payload
         .map(p => s"""--data '$p'""")
         .getOrElse("")} $url"""
+    }
   }
 
   object RestQuery {
@@ -187,10 +199,35 @@ object RestQueryController {
   val RestQueryForm: Form[RestQuery] = Form[RestQuery](
     mapping(
       "method" -> text,
-      "url"    -> text,
+      "url"    -> text.verifying("error.url", u => Try(new URL(u)).isSuccess),
       "payload" -> optional(text)
         .verifying(validJson)
-        .transform[Option[JsValue]](_.map(Json.parse), _.map(Json.prettyPrint))
+        .transform[Option[JsValue]](_.map(Json.parse), _.map(Json.prettyPrint)),
+      "headers" -> optional(text).verifying("error.headers", _.forall{ h => parseHeaders(h); true})
     )(RestQuery.apply)(RestQuery.unapply))
+
+  def prettyPrintHeaders(headers: Seq[(String,String)]): Option[String] = 
+    if (headers.isEmpty) None 
+    else 
+      Some( headers.map { 
+        case (key, value) => s"$key: $value"
+      }.mkString("\n"))
+
+  def parseHeaders(headers: String): Map[String,String] = 
+    if (headers.isEmpty) Map.empty 
+    else 
+      headers
+        .lines
+        .map { line => 
+          val key = line.takeWhile(_ != ':')
+          if(key.nonEmpty) {
+            val value = line.drop(key.length + 1)
+            if(value.nonEmpty) {
+              Some((key.trim(), value.trim()))
+            } else None
+          } else None
+        }
+        .collect {case Some(x) => x}
+        .toMap
 
 }
